Start merging textMesh with labelMesh
